Combined app files, remove redundant files and created initial registration integration tests
Combined course_app and registration_app to app.py. Removed unnecessary files. Created registration_integration_tests
